refactor: add useAnimation for reusability
